requirements.txt
httpx
pyyaml
beautifulsoup4
